+++ cache/api/CHANGELOG.md
@@ -1,5 +1,12 @@
 # @pnpm/cache.api
 
+## 1000.0.9
+
+### Patch Changes
+
+- Updated dependencies [fee898f]
+  - @pnpm/config@1002.3.1
+
 ## 1000.0.8
 
 ### Patch Changes