+++ pkg-manager/plugin-commands-installation/CHANGELOG.md
@@ -1,5 +1,24 @@
 # @pnpm/plugin-commands-installation
 
+## 1001.5.1
+
+### Patch Changes
+
+- 546ab37: Throws an error when the value provided by the `--allow-build` option overlaps with the `pnpm.ignoredBuildDependencies` list [#9105](https://github.com/pnpm/pnpm/pull/9105).
+- Updated dependencies [fee898f]
+  - @pnpm/config@1002.3.1
+  - @pnpm/cli-utils@1000.0.10
+  - @pnpm/deps.status@1001.2.2
+  - @pnpm/plugin-commands-env@1000.0.10
+  - @pnpm/plugin-commands-rebuild@1001.1.8
+  - @pnpm/store-connection-manager@1000.0.9
+  - @pnpm/workspace.state@1001.1.5
+  - @pnpm/core@1003.0.2
+  - @pnpm/workspace.find-packages@1000.0.10
+  - @pnpm/get-context@1001.0.5
+  - @pnpm/outdated@1001.0.7
+  - @pnpm/filter-workspace-packages@1000.0.10
+
 ## 1001.5.0
 
 ### Minor Changes