+++ releasing/plugin-commands-deploy/CHANGELOG.md
@@ -1,5 +1,16 @@
 # @pnpm/plugin-commands-deploy
 
+## 1001.1.8
+
+### Patch Changes
+
+- Updated dependencies [fee898f]
+- Updated dependencies [546ab37]
+  - @pnpm/config@1002.3.1
+  - @pnpm/plugin-commands-installation@1001.5.1
+  - @pnpm/cli-utils@1000.0.10
+  - @pnpm/lockfile.fs@1001.1.3
+
 ## 1001.1.7
 
 ### Patch Changes