+++ releasing/plugin-commands-publishing/CHANGELOG.md
@@ -1,5 +1,14 @@
 # @pnpm/plugin-commands-publishing
 
+## 1000.1.8
+
+### Patch Changes
+
+- Updated dependencies [fee898f]
+  - @pnpm/config@1002.3.1
+  - @pnpm/cli-utils@1000.0.10
+  - @pnpm/plugin-commands-env@1000.0.10
+
 ## 1000.1.7
 
 ### Patch Changes