+++ reviewing/dependencies-hierarchy/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @pnpm/reviewing.dependencies-hierarchy
 
+## 1001.0.5
+
+### Patch Changes
+
+- @pnpm/lockfile.fs@1001.1.3
+
 ## 1001.0.4
 
 ### Patch Changes