+++ store/plugin-commands-server/CHANGELOG.md
@@ -1,5 +1,14 @@
 # @pnpm/plugin-commands-server
 
+## 1000.0.10
+
+### Patch Changes
+
+- Updated dependencies [fee898f]
+  - @pnpm/config@1002.3.1
+  - @pnpm/cli-utils@1000.0.10
+  - @pnpm/store-connection-manager@1000.0.9
+
 ## 1000.0.9
 
 ### Patch Changes