+++ store/store-connection-manager/CHANGELOG.md
@@ -1,5 +1,12 @@
 # @pnpm/store-connection-manager
 
+## 1000.0.9
+
+### Patch Changes
+
+- Updated dependencies [fee898f]
+  - @pnpm/config@1002.3.1
+
 ## 1000.0.8
 
 ### Patch Changes