+++ workspace/find-packages/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @pnpm/find-workspace-packages
 
+## 1000.0.10
+
+### Patch Changes
+
+- @pnpm/cli-utils@1000.0.10
+
 ## 1000.0.9
 
 ### Patch Changes